Add kernel tweak to post-creation hook
Jupyter launches a kernel in order to 'run' a notebook. The configuration of
that kernel is located, generally, where `jupyter --data-dir` is, under the
`/kernels` subfolder. By default, the kernel launches the python binary found at
`pipenv --venv`; this modification adds a `cd` command to run before the python
binary, which makes sure that the python binary is launched from the project
root (here, the output of `git rev-parse --show-toplevel`.

If git is not initialized, the `cd` step fails and the python binary is launched
from the notebook location.

I don't know what happens if pipenv is not installed; I'm guessing the binary
won't be found, and the kernel will fail to start. Possible fix to that could be
to add a conditional that fails over to some systemwide Python install, but I
think we are expecting `pipenv` in other parts of this project template
anyway (as in issue #1).